feat(api): Move parameters to Model
