+++ examples/api/program-management/app.ts
@@ -1,5 +1,5 @@
 import {getRandom} from '@luma.gl/api';
-import {setDeviceParameters, setParameters} from '@luma.gl/webgl';
+import {setParameters} from '@luma.gl/webgl';
 import {dirlight as dirlightBase} from '@luma.gl/shadertools';
 import {RenderLoop, Model, ProgramManager, AnimationProps, CubeGeometry} from '@luma.gl/engine';
 import {Matrix4, radians} from '@math.gl/core';
@@ -62,10 +62,6 @@ export default class AppRenderLoop extends RenderLoop {
   constructor({device, gl, aspect}: AnimationProps) {
     super();
 
-    setDeviceParameters(device, {
-      depthWriteEnabled: true,
-      depthCompare: 'less-equal'
-    });
     setParameters(gl, {
       clearColor: [0, 0, 0, 1],
       clearDepth: 1
@@ -115,6 +111,10 @@ export default class AppRenderLoop extends RenderLoop {
               eye: [0, 0, -8]
             }),
             uColor: colors[i]
+          },
+          parameters: {
+            depthWriteEnabled: true,
+            depthCompare: 'less-equal'
           }
         })
       };
