+++ examples/showcase/gltf/app.ts
@@ -4,7 +4,7 @@ import {Device} from '@luma.gl/api';
 import {RenderLoop, AnimationProps, Timeline} from '@luma.gl/engine';
 import {createGLTFObjects, GLTFEnvironment, VRDisplay} from '@luma.gl/experimental';
 import GL from '@luma.gl/constants';
-import {clear, log, lumaStats, setDeviceParameters} from '@luma.gl/webgl';
+import {clear, log, lumaStats} from '@luma.gl/webgl';
 import {parse} from '@loaders.gl/core';
 import {GLTFLoader} from '@loaders.gl/gltf';
 import {Matrix4, radians} from '@math.gl/core';
@@ -231,12 +231,6 @@ export default class AppRenderLoop extends RenderLoop {
 
     // @ts-ignore
     // this._setDisplay(new VRDisplay());
-
- onInitialize({gl, canvas}) {
-    setDeviceParameters(gl, {
-      depthWriteEnabled: true,
-      depthCompare: 'less-equal'
-    });
 
     this.environment = new GLTFEnvironment(device, {
       brdfLutUrl: `${GLTF_BASE_URL}/brdfLUT.png`,
