+++ modules/shadertools/test/gpu-test-utils.js
@@ -39,10 +39,10 @@ function glErrorShouldBe(gl, glErrors, opt_msg) {
 export function initializeGL(canvas) {
   const gl = createTestContext(canvas);
   setParameters(gl, {
-    viewport: [0, 0, canvas.width, canvas.height],
+    viewport: [0, 0, canvas.width, canvas.height]
   });
   setParameters(gl, {
-  clearColor: [0, 0, 0, 1],
+    clearColor: [0, 0, 0, 1],
     clearDepth: 1
   });
   gl.clear(gl.COLOR_BUFFER_BIT | gl.DEPTH_BUFFER_BIT);
