+++ modules/webgl/test/classes/texture.spec.js
@@ -1,5 +1,6 @@
 /* eslint-disable max-len */
 import test from 'tape-promise/tape';
+import {fixture} from 'test/setup';
 
 import GL from '@luma.gl/constants';
 import {isWebGL2} from '@luma.gl/webgl';
@@ -7,7 +8,61 @@ import {Buffer, Texture2D, getKey, readPixelsToArray} from '@luma.gl/webgl';
 
 import {TEXTURE_FORMATS} from '@luma.gl/webgl/classes/texture-formats';
 
-import {fixture} from 'test/setup';
+export const SAMPLER_PARAMETERS = {
+  [GL.TEXTURE_MIN_FILTER]: {
+    [GL.LINEAR]: 'interpolated texel',
+    [GL.NEAREST]: 'nearest texel',
+    [GL.NEAREST_MIPMAP_NEAREST]: 'nearest texel in closest mipmap',
+    [GL.LINEAR_MIPMAP_NEAREST]: 'interpolated texel in closest mipmap',
+    [GL.NEAREST_MIPMAP_LINEAR]: 'average texel from two closest mipmaps',
+    [GL.LINEAR_MIPMAP_LINEAR]: 'interpolated texel from two closest mipmaps'
+  },
+
+  [GL.TEXTURE_MAG_FILTER]: {
+    [GL.LINEAR]: 'interpolated texel',
+    [GL.NEAREST]: 'nearest texel'
+  },
+
+  [GL.TEXTURE_WRAP_S]: {
+    [GL.REPEAT]: 'use fractional part of texture coordinates',
+    [GL.CLAMP_TO_EDGE]: 'clamp texture coordinates',
+    [GL.MIRRORED_REPEAT]:
+      'use fractional part of texture coordinate if integer part is odd, otherwise `1 - frac'
+  },
+
+  [GL.TEXTURE_WRAP_T]: {
+    [GL.REPEAT]: 'use fractional part of texture coordinates',
+    [GL.CLAMP_TO_EDGE]: 'clamp texture coordinates',
+    [GL.MIRRORED_REPEAT]:
+      'use fractional part of texture coordinate if integer part is odd, otherwise `1 - frac'
+  }
+};
+
+export const SAMPLER_PARAMETERS_WEBGL2 = {
+  [GL.TEXTURE_WRAP_R]: {
+    [GL.REPEAT]: 'use fractional part of texture coordinates',
+    [GL.CLAMP_TO_EDGE]: 'clamp texture coordinates',
+    [GL.MIRRORED_REPEAT]:
+      'use fractional part of texture coordinate if integer part is odd, otherwise `1 - frac'
+  },
+
+  [GL.TEXTURE_COMPARE_MODE]: {
+    [GL.NONE]: 'no comparison of `r` coordinate is performed',
+    [GL.COMPARE_REF_TO_TEXTURE]:
+      'interpolated and clamped `r` texture coordinate is compared to currently bound depth texture, result is assigned to the red channel'
+  },
+
+  [GL.TEXTURE_COMPARE_FUNC]: {
+    [GL.LEQUAL]: 'result = 1.0 0.0, r <= D t r > D t',
+    [GL.GEQUAL]: 'result = 1.0 0.0, r >= D t r < D t',
+    [GL.LESS]: 'result = 1.0 0.0, r < D t r >= D t',
+    [GL.GREATER]: 'result = 1.0 0.0, r > D t r <= D t',
+    [GL.EQUAL]: 'result = 1.0 0.0, r = D t r ≠ D t',
+    [GL.NOTEQUAL]: 'result = 1.0 0.0, r ≠ D t r = D t',
+    [GL.ALWAYS]: 'result = 1.0',
+    [GL.NEVER]: 'result = 0.0'
+  }
+};
 
 test('WebGL#Texture2D construct/delete', (t) => {
   const {gl} = fixture;
@@ -651,3 +706,24 @@ test('WebGL2#Texture2D generateMipmap', (t) => {
 
   t.end();
 });
+
+// Shared with texture*.spec.js
+export function testSamplerParameters({t, texture, parameters}) {
+  for (const parameterName in parameters) {
+    const values = parameters[parameterName];
+    const parameter = Number(parameterName);
+    for (const valueName in values) {
+      const value = Number(valueName);
+      texture.setParameters({
+        [parameter]: value
+      });
+      const name = texture.constructor.name;
+      const newValue = texture.getParameter(parameter);
+      t.equals(
+        newValue,
+        value,
+        `${name}.setParameters({[${getKey(GL, parameter)}]: ${getKey(GL, value)}}) read back OK`
+      );
+    }
+  }
+}
